BackEnd/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pg8000
pydantic[email]
boto3
opencv-python
python-jose[cryptography]
passlib[bcrypt]
python-multipart
slowapi
